pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vas"]

[tool.hatch.version]
path = "src/vas/__init__.py"

[project]
name = "voice-segmentation"
version = "0.1.0"
description = "A project that holds many strategies for segmenting audio files in speech pauses. "
readme = "README.md"
license = "MIT"
authors = [{ name = "Andre Koraleski", email = "andrekorale@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "pydantic-settings>=2.13.1",
    "soundfile>=0.13.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]


# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
src = ["src"]
exclude = ["tests"]

[tool.ruff.lint]
select = [
  "D",
  "E",
  "W",
  "F",
  "I",
  "B",
  "C4",
  "UP",
  "N",
  "SIM",
  "RUF",
  "S",
  "PTH",
  "ERA",
  "PL",
  "TRY",
  "ANN",
  "TC",
  "ISC",
  "PIE",
]
ignore = ["D203", "D213"]
unfixable = ["ERA001", "F401"]
per-file-ignores = { "tests/**/*.py" = ["S101", "ANN", "PLR2004"] }
isort = { known-first-party = ["vas"] }
pylint = { max-args = 8 }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
exclude = ["tests/"]
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_subclassing_any = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: Requires a simulation server started in-process (slower)",
]
addopts = [
  "--strict-markers",
  "--strict-config",
  "-ra",
  "--cov=vas",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--cov-branch",
]
xfail_strict = true

[tool.coverage.run]
branch = true
source = ["src/vas"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@(abc\\.)?abstractmethod",
]
